requirements.txt
Pillow>=8.1.1
einops == 0.3.0
imageio == 2.9.0
matplotlib == 3.3.2
numpy == 1.19.2
opencv_python == 3.4.2.17
scikit_image == 0.17.2
scipy == 1.5.2
sklearn
torch == 1.7.0
torchvision == 0.8.1
tqdm == 4.51.0
transformers == 3.5.1
utils == 1.0.1
captum
wget
ftfy
regex
